// commercial/template/src/utility/enum/queryTag.enum.ts
/**
 * @Author         : David Nguyễn <davidnguyen67dev@gmail.com>
 * @CreatedDate    : 2024-06-29 19:55:00
 * @LastEditors    : David Nguyễn <davidnguyen67dev@gmail.com>
 * @LastEditDate   : 2024-06-30 11:47:05
 * @FilePath       : queryTag.enum.ts
 * @CopyRight      : Con chù chù 🥴🥴
 **/

export enum QUERY_TAG {
  USER = 'USER',
  COUNT_USER = 'COUNT_USER',
}